+++ packages/components/src/Menu/MenuItem/index.ts
@@ -25,4 +25,3 @@
  */
 
 export * from './MenuItem'
-export * from './menuItemCustomization'