+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ComboboxMultiOption` prop `hideCheckMark` (instead use `indicator={false}`)
 - `psuedo` support from design-tokens (prefer using `styled` when psuedo selectors are required)
   - Includes changes to `Box` & `IconButton`
+- `MenuList`, `MenuGroup` & `MenuItem` no longer supports `customizationProps`
 
 ## [0.8.7]